Add an event bus
Convert everything to being event sourced as much as possible. Make a comprehensive set of events for invocations.

Write the events as they are published, and expand the testing to cover several layers of abstraction, including synchronous routines, async routines, async functions that can be used in routines, async routines that call async routines, and returning values from awaited routines.

Fix a bug in the implementation of Invocation.__await__ that stopped return values from working correctly.

Stub out handling errors, but throw a runtime error if we ever encounter one.