Make our own threading-based executor
ThreadPoolExecutor uses daemon threads that don't shut down the way that I expect them when things are force shut down. At this point I'm not too concerned about pooling threads, so I skipped it for simplicity. As it's written now it takes 3 Ctrl-C to really shut down versus 2 normally for the interpreter, but that's better than it never finishing.