Split up the main function into several actors
The queuer takes invocations submitted events and puts them on the queue.

The runner takes invocations or continuations from the queue and creates a future for them.

The waiter waits for the futures to complete and writes events for how they completed.

The continuer watches for completed invocations and continues any suspended invocation that called to them.

The inspector watches for all events and writes them out to stdout.