Allow for a graceful shutdown
It doesn't handle waiting for an async function to complete, but it waits for each individual segment of the async function to complete or a sync function to complete before shutting everything down.

Notably, this required two semaphores, one to guard reservations, and another to guard the actual running, or else it could hang if the shutdown happened after the concurrency was reserved but before there was a task on the queue to run.