Use RabbitMQ for invocation queueing
This one was a beast to implement, both because it was complex and because it uncovered a number of bugs, and required extensive reworking to make everything worked with serialization of the routines and arguments.

As is the mode of this development, this is the simplest thing that works. It supports only one statically-defined queue. The routine registry is nothing more than a global dictionary, and the routines are ensured to be present in the registry when they are serialized. To avoid issues where the queue doesn't exist or where the routines don't exist during deserialization,  the queue is declared and purged at startup, before the first messages are enqueued.

Serialization and deserialization use UTF-8 encoded JSON. `id`, `args`, and `kwargs` are expected to be JSON-serializable, and only the routine has a unserializable data type that requires the routine registry.

It turns out that the implementation of the waiter actors was a busy-wait loop under most circumstances. The only way that I saw to fix that was to have the runner thread itself handle acknowledgements and making the events from the results of running the routines.

Because waiter is no longer watching the tasks to complete because they finalize their own events, we no longer need the `InvocationStarted` and InvocationResumed` events to have any reference to the future or generator associated with the invocation.

Because of serialization, the existing equality implementation is not sufficient. Instead, we need to compare the ids of the invocations, which are unique and random and are a good enough signal for equality. Eventually we may have a good solution in the Invocation data type, but for now whenever the invocation is used as a dictionary key, we need to use its id instead.

`ruff check` saved me from a subtle bug, where I was using lambda functions to start threads in a loop, but because the loop variables can change the lambda could potentially be using the wrong values. Instead, I needed to use `partial` to capture the values of the loop.

Not withstanding that, using `lambda` functions for the main function still felt better. Those aren't in a loop, so don't have the same worries about the variable changing value before it is read.

Added a `Consumer` abstraction instead of a simple shared queue. This allows for acknowledgements as appropriate for RabbitMQ without making things more complex for the queuer. The consumer is an iterator, so when it gets shut down, it will finish producing new messages.

In chasing down one particular rabbit hole, I updated the `Concurrency` class so that multiple threads would be able to shut down and wait for the concurrency to finish, by adding a wait lock so that threads would do the waiting process one at a time.

I was able to implement the consumer methods in such a way that they would be blocking calls, which was more difficult because pika is not thread-safe, except for one helper method that can enqueue a callback that is automatically executed when waiting for the next message from the consumer.

It's important that the consumer not be shut down until all the active messages have been processed. This avoids particularly an issue where the consumer is not active, and thus not communicating with RabbitMQ, while the task is still running. Instead, the concurrency needs to be shut down and cleared out first.

Include a counter when submitting tasks to the executor, to make it a little easier to track down thread output.