Avoid a separate queuer actor
When this library is put into an application in place of celery, there's no good place for an enqueuer actor to live, so it needs to just be part of the producer. Do the same for the continuation queuer, for consistency.

Doing this work surfaced an issue where messages could be published before the consuming actors were subscribed. To avoid this, the event consuming queues are created at startup and passed into the actors, ensuring that the subscriptions are aware of all relevent events.