Calling a routine should give a suspension
Calling a routine will the normal way to create the object that will be awaited in qio async routines. I'm working under the theory that returning the suspension directly, and then having __await__ yield self, will be a better API design than the current. Time will tell as this is updated to have other types of suspensions, whether this API remains optimal.