Refactor the bus to make a transport
A transport is the interface to the remote backend for the bus. It doesn't currently support filtering, although that could theoretically be added later. It doesn't know anything about types, and doesn't multiplex messages to subscribers. It interfaces with a remote transport so that events on other hosts are received.