Clean up the Broker API
Create a new Message class, that contains the raw message body and nothing else. It doesn't handle serialization or deserialization. It's a custom class so that brokers can use the instances to keep track of the messages while they are being processed and communicate it to the broker as needed.

Pull the serialization and deserialization out of the broker and make it a concern of other parts of Qio.