Ack on suspend with pika
AMQP 0.9.1 doesn't support changing the prefetch window on a subscription (what connects a channel to a queue). AMQP 1.0 supports it, but that takes a different library to resolve. With pika it seems the best we can do is to ack the message on suspend so that we don't block further activity from happening. This means that if a suspended worker is shut down, or it is shut down after the task resumes but before it completes, it will not be re-queued.