Create an abstract interface for suspensions
Notably, this required having the qio class, rather than the invocation or invocation suspension, own submitting the suspension. This required putting a contextvar global handler to enqueue invocations, since invocations are a dependency of the Qio interface so the Qio interface can't be a dependency of invocation suspension.

Uses futures as the common abstraction due to the ability to do multiple waiting. And extend the custom Thread class to handle copying contextvar context.