Make Suspendables Suspensions
Suspendable is a nearly useless abstraction over Awaitable. It's good for the type checker in some cases, but we still have to support regular awaitables, so it really doesn't help us practically.

But a common pattern has emerged in places that need to return a Suspendable, namely, that I'm creating a generator that only returns a single, specialized, Suspension. This has happened now for Invocation, Sleep, and Gather. To to simplify the `suspending` decorator so that it only returns a single suspension, I've made those types Suspendable, and changed Suspendable to inherit also from Suspension, given a default implementation of __await__ that is perfect for those items, and changed the `suspending` decorator to make the wrapped function return only a single Suspendable (called, aptly, `Suspending`). Something eels a bit double-wrapped about that, but I'll figure that out in another iteration.