Add a select protocol and queues
The motivating reason for this is to allow waiting for getting values from multiple queues simultaneously, taking only one value from any queue.

The primary source of inspiration for this came from the implementation of Go channels described by Dmitry Vykov in the paper titled "Go channels on steriods" (https://docs.google.com/document/d/1yIAYmbvL3JxOKOjuCyon7JhW4cSv1wy5hC0ApeGMV9s/pub).

However, the implementation diverges very significantly. I gleaned the critical hint that the select call had a shared lock between all waiters in order to ensure that only one waiter would be able to interact with the channel. To handle this, I use a context manager that wraps the API needed to send the result to the select.

Rather than intricately tying the select to a specific data type, this introduces a protocol that I've created initial implementations of three different types of things that remind me a bit of Go's channels. The one that is most different, and most interesting, is SwapQueue, where two sides synchronously swap values, so it acts like a synchronous duplex pipe.

The key insight is that, in Python, we can send None as the type when we're receiving nothing (much as returning nothing is actually returning None), and that we can also send None when we're just getting a value. This means that the select protocol doesn't need to be aware of whether a given select action is sending a value or receiving it.

I need this in order to implement fair waiting concurrency for multiple queues in the StubBroker.