Enable multiple queues in pika
At least for RabbitMQ, the messages are delivered less fairly than the stub broker, but it doesn't seem to starve any queue out either. The prefetch count seems to be initially filled from a single queue, but things can be fairly distributed after that. While RabbitMQ truly reflects that there are multiple consumers to a queue on the same channel, it's also not consistent about giving additional priority to queues with more consumers. Because the protocol allows it, I think it's best to think of these as RabbitMQ-specific implementation details.

With the general fairness, I was able to modify the test so that the message count was larger and it could take a more holistic look at the entirety of the message consumption, rather than being overly biased by the initial distribution of messages. For the weighted behavior, I wasn't able to determine a sufficiently reliable way to test that it gives that behavior in some normal circumstances, so I've skipped that test for the Pika broker.