Rework event and instance id naming pattern
Since Invocation is a Suspension, it's becoming clear that we can have lifecycle events for all suspensions. But "suspension_id" being the id just because everything is a suspension, even when the type is more specialized, seems suboptimal.

Instead, since the event_id is much less likely to be accessed, it can have the name "event_id", leaving the "id" to be the instance id, or aggregate id in event sourcing. This has the benefit of avoiding the "aggregate" verbage, which IMO is not very approachable as a concept. I think it's a little unintuitive that event.id will be the id of the object of the event rather than the event, but I think it's outweighed by the benefits of doing it this way.