Add a package-level activate context manager
I believe that this has more staying power than the existing API. It could be that even this can eventually be eliminated, but I'm not sure about that just yet. Either way, this avoids exposing QueueIO as a public interface. Which is good, because I'm coming to think that the functions it provides probably all would better live in other places.