Rethink the CLI commands
There are two main sections. One for doing the main operations, monitoring and running invocations, and the other for describing command related to entities.

Create a new sync command that ensures the queues are on the broker, and don't auto-create queues. A future iteration will allow for creating a lockfile and also be able to drop unneeded queues from the broker.